+++ NEWS.md
@@ -3,7 +3,7 @@
 -------------------------------
 
 # Version 0.8.5
-- New functions to convert SpatialPixelsMeteorology and SpatialGridMeteorology into objects of packages 'spacetime' and 'stars'
+- New functions to convert SpatialPointsMeteorology, SpatialPixelsMeteorology and SpatialGridMeteorology into objects of packages 'spacetime' and 'stars'
 
 # Version 0.8.4
 - New function to calculate rainfall erosivity